Fix CI: unpin mypy python_version, drop an ill-posed test, pin LLM dtype
CI failed on mypy and one test. Both causes were the local environment being
far behind what CI resolves (numpy 1.24/scikit-learn 1.8 locally against
numpy 2.4/scikit-learn 1.9/pandas 3.0 in CI), so the fixes were verified in a
clean venv installed the way CI installs.

mypy
- Unpin `python_version`. Pinning it below the interpreter mypy runs on makes
  numpy 2.4's stubs unparseable: they use PEP 695 `type` statements, which
  mypy rejects outright when told to assume 3.11. Ruff's target-version
  already covers syntax compatibility and the test matrix really runs on 3.11.
- Add `torch.*` and `dotenv.*` to ignore_missing_imports. The lint job installs
  only [dev], so both are genuinely absent there; the imports are lazy and
  guarded by an ImportError that names the extra.
- Drop the phate/pacmap/sentencepiece overrides. Those are loaded through
  __import__ or not imported at all, so mypy never sees them and reported the
  overrides as unused.

Ill-posed test
- `test_equidistant_points_cannot_be_preserved` asserted trustworthiness < 0.5
  for `np.eye(20)`. Every pairwise distance in that input is sqrt(2), so the
  k-neighbourhoods are decided purely by how the sort breaks ties:
  scikit-learn 1.8 returns 0.27 and 1.9 returns 0.95 for the identical call.
  The metric is undefined for this input rather than low. An earlier version of
  this test asserted > 0.8 and a later one < 0.5, and each passed on exactly
  one of those releases. It now asserts only the guaranteed invariant -- finite
  and in range -- with the meaningful behaviour covered by the neighbouring
  tests, which use well-separated data.

LLMEmbedder dtype
- Reproducing CI's stubs surfaced a real bug. Dividing the float32 embedding by
  float64 norms promoted the result, so the embedder returned float32 normally
  and float64 whenever `output_dim` was set -- contradicting its docstring and
  doubling the size of every cached .npy for that configuration. Cast back
  explicitly.

Docs
- Record both traps in CLAUDE.md and CONTRIBUTING.md, with the venv recipe for
  checking a change against the versions CI will actually resolve.

Verified in a venv matching CI exactly (numpy 2.4.6, scikit-learn 1.9.1,
scipy 1.17.1, pandas 3.0.5): ruff, ruff format and mypy clean, 411 tests pass,
and the smoke job's own assertions pass against the real console scripts,
including the bit-identical rerun check.

+++ src/prot2vec/embedders/llm.py
@@ -257,8 +257,12 @@ class LLMEmbedder(SequenceEmbedder):
         if self.output_dim is not None:
             # Gemini normalises only its full-width output; truncated vectors
             # must be re-normalised before cosine distances mean anything.
+            # The explicit cast back matters: dividing float32 by float64 norms
+            # promotes the result, which would make the returned dtype -- and
+            # so the size of every cached .npy -- depend on whether output_dim
+            # happened to be set.
             norms = np.linalg.norm(result, axis=1, keepdims=True)
-            result = result / np.where(norms == 0, 1.0, norms)
+            result = (result / np.where(norms == 0, 1.0, norms)).astype(np.float32)
         return result
 
     # ------------------------------------------------------------------
